Update RNReactLoggingModule.java
Change long to double

+++ android/src/main/java/com/reactlibrary/RNReactLoggingModule.java
@@ -29,7 +29,7 @@ public class RNReactLoggingModule extends ReactContextBaseJavaModule {
     private String tag = "RNReactLogging";
     private boolean consoleLog = true;
     private boolean fileLog = false;
-    private long maxFileSize = 512 * 1024; // 512 kb
+    private double maxFileSize = 512 * 1024; // 512 kb
 
     public RNReactLoggingModule(ReactApplicationContext reactContext) {
         super(reactContext);
@@ -149,7 +149,7 @@ public class RNReactLoggingModule extends ReactContextBaseJavaModule {
     }
 
     @ReactMethod
-    public void setMaxFileSize(long maxFileSize) {
+    public void setMaxFileSize(double maxFileSize) {
         this.maxFileSize = maxFileSize;
     }
